Move the gbdt and vm_forest back to the eval module.
