+++ vespalib/src/tests/eval/gbdt/gbdt_benchmark.cpp
@@ -1,7 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #include <vespa/vespalib/testkit/test_kit.h>
-#include <vespa/vespalib/eval/llvm/gbdt.h>
-#include <vespa/vespalib/eval/llvm/vm_forest.h>
+#include <vespa/vespalib/eval/gbdt.h>
+#include <vespa/vespalib/eval/vm_forest.h>
 #include <vespa/vespalib/eval/llvm/deinline_forest.h>
 #include <vespa/vespalib/eval/llvm/compiled_function.h>
 #include <vespa/vespalib/eval/function.h>
