+++ vespalib/src/vespa/vespalib/eval/llvm/deinline_forest.cpp
@@ -1,7 +1,6 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/fastos/fastos.h>
-#include "gbdt.h"
 #include "deinline_forest.h"
 
 namespace vespalib {
