+++ vespalib/src/vespa/vespalib/eval/llvm/llvm_wrapper.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include <vespa/vespalib/eval/function.h>
-#include "gbdt.h"
+#include <vespa/vespalib/eval/gbdt.h>
 
 #include <llvm/IR/LLVMContext.h>
 #include <llvm/IR/Module.h>
