+++ vespalib/src/vespa/vespalib/eval/vm_forest.cpp
